Changed socketctl to use a PyBluez BluetoothSocket instead of a native Python socket. Also generally refactored the codebase, moving all code into a separate package and added a setup.py. However, I can't get the server to work on my Windows laptop yet. I would like to know if it still works on Linux with these changes.
