+++ scoutingserver/__init__.py
@@ -1,0 +1,5 @@
+from scoutingserver.server import Server
+
+def main():
+    server = Server()
+    server.main()
