+++ scoutingserver/dataconstants.py
