+++ scoutingserver/interface/printing.py
@@ -1,4 +1,4 @@
-from interface.logger import log as log_
+from scoutingserver.interface.logger import log as log_
 
 BASE = '\033[9'
 BASE_END = 'm'
